staffrole增删查改
role增删查改

+++ backend/src/main/java/com/xietong/model/entity/SecurityStaffRoleDO.java
@@ -14,7 +14,7 @@ import lombok.RequiredArgsConstructor;
 @TableName("security_staff_role")
 public class SecurityStaffRoleDO {
     @TableId
-    private int id;
+    private Long id;
     private String staffId;
-    private int roleId;
+    private Long roleId;
 }
